made_team/app_project/requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
login==0.0.6
MarkupSafe==2.1.2
mysqlclient==2.1.1
SQLAlchemy==2.0.13
typing_extensions==4.5.0
Werkzeug==2.3.4
